Move animation functionality into a mixin, to lighten up QuestionnaireBodyCreate
